requirements.txt
django
djangorestframework
python-decouple
sqlparse
whitenoise
asgiref
dj-database-url
python-webpack-boilerplate
Pillow
drf-yasg[validation]
